+++ bin/handlebarsc.c
@@ -39,6 +39,7 @@
 #include "handlebars_opcodes.h"
 #include "handlebars_opcode_printer.h"
 #include "handlebars_opcode_serializer.h"
+#include "handlebars_partial_loader.h"
 #include "handlebars_string.h"
 #include "handlebars_token.h"
 #include "handlebars_utils.h"
@@ -58,7 +59,10 @@ char * input_name = NULL;
 char * input_data_name = NULL;
 char * input_buf = NULL;
 size_t input_buf_length = 0;
+char * partial_path = ".";
+char * partial_extension = ".hbs";
 unsigned long compiler_flags = 0;
+short enable_partial_loader = 0;
 
 enum handlebarsc_mode {
     handlebarsc_mode_usage = 0,
@@ -70,7 +74,7 @@ enum handlebarsc_mode {
     handlebarsc_mode_debug
 };
 
-static enum handlebarsc_mode mode;
+static enum handlebarsc_mode mode = handlebarsc_mode_execute;
 
 /**
  * http://linux.die.net/man/3/getopt_long
@@ -95,21 +99,17 @@ static void readOpts(int argc, char * argv[])
         {"template",  required_argument,        0,  't' },
         {"data",      required_argument,        0,  'D' },
         // compiler flags
-        {"compat",    no_argument,              0,  'C' },
-        {"known-helpers-only", no_argument,     0,  'K' },
-        {"string-params", no_argument,          0,  'S' },
-        {"track-ids", no_argument,              0,  'T' },
-        {"no-escape", no_argument,              0,  'U' },
-        {"ignore-standalone", no_argument,      0,  'G' },
-        {"alternate-decorators", no_argument,   0,  'A' },
-        {"strict",    no_argument,              0,  's' },
-        {"assume-objects", no_argument,         0,  'a' },
-        {"mustache-style-lambdas", no_argument, 0,  'L' },
+        {"flags",     required_argument,        0,  'F' },
+        // loaders
+        {"partial-loader", no_argument,         0,  'P' },
+        {"partial-path",   required_argument,   0,  'A' },
+        {"partial-ext",    required_argument,   0,  'X' },
+        // end
         {0,           0,                        0,  0   }
     };
 
 start:
-    c = getopt_long(argc, argv, "hlpcVt:", long_options, &option_index);
+    c = getopt_long(argc, argv, "hlpcVt:D:", long_options, &option_index);
     if( c == -1 ) {
         return;
     }
@@ -139,35 +139,48 @@ start:
             break;
 
         // compiler flags
-        case 'C':
-            compiler_flags |= handlebars_compiler_flag_compat;
+        case 'F':
+            // we could do this more efficiently
+            if (NULL != strstr(optarg, "compat")) {
+                compiler_flags |= handlebars_compiler_flag_compat;
+            }
+            if (NULL != strstr(optarg, "known_helpers_only")) {
+                compiler_flags |= handlebars_compiler_flag_known_helpers_only;
+            }
+            if (NULL != strstr(optarg, "string_params")) {
+                compiler_flags |= handlebars_compiler_flag_string_params;
+            }
+            if (NULL != strstr(optarg, "track_ids")) {
+                compiler_flags |= handlebars_compiler_flag_track_ids;
+            }
+            if (NULL != strstr(optarg, "no_escape")) {
+                compiler_flags |= handlebars_compiler_flag_no_escape;
+            }
+            if (NULL != strstr(optarg, "ignore_standalone")) {
+                compiler_flags |= handlebars_compiler_flag_ignore_standalone;
+            }
+            if (NULL != strstr(optarg, "alternate_decorators")) {
+                compiler_flags |= handlebars_compiler_flag_alternate_decorators;
+            }
+            if (NULL != strstr(optarg, "strict")) {
+                compiler_flags |= handlebars_compiler_flag_strict;
+            }
+            if (NULL != strstr(optarg, "assume_objects")) {
+                compiler_flags |= handlebars_compiler_flag_assume_objects;
+            }
+            if (NULL != strstr(optarg, "mustache_style_lambdas")) {
+                compiler_flags |= handlebars_compiler_flag_mustache_style_lambdas;
+            }
             break;
-        case 'K':
-            compiler_flags |= handlebars_compiler_flag_known_helpers_only;
-            break;
-        case 'S':
-            compiler_flags |= handlebars_compiler_flag_string_params;
-            break;
-        case 'T':
-            compiler_flags |= handlebars_compiler_flag_track_ids;
-            break;
-        case 'U':
-            compiler_flags |= handlebars_compiler_flag_no_escape;
-            break;
-        case 'G':
-            compiler_flags |= handlebars_compiler_flag_ignore_standalone;
+
+        case 'P':
+            enable_partial_loader = 1;
             break;
         case 'A':
-            compiler_flags |= handlebars_compiler_flag_alternate_decorators;
+            partial_path = optarg;
             break;
-        case 's':
-            compiler_flags |= handlebars_compiler_flag_strict;
-            break;
-        case 'a':
-            compiler_flags |= handlebars_compiler_flag_assume_objects;
-            break;
-        case 'L':
-            compiler_flags |= handlebars_compiler_flag_mustache_style_lambdas;
+        case 'X':
+            partial_extension = optarg;
             break;
 
         // input
@@ -236,13 +249,50 @@ static void readInput(void)
 
 static int do_usage(void)
 {
-    fprintf(stderr, "Usage: handlebarsc [--lex|--parse|--compile|--execute] [--data JSON_FILE] [TEMPLATE]\n");
+    fprintf(stdout,
+        "Usage: handlebarsc [OPTIONS]\n"
+        "Example: handlebarsc -t foo.hbs -D bar.json\n"
+        "\n"
+        "Mode options:\n"
+        "  -h, --help            Show this message\n"
+        "  -e, --execute         Execute the specified template (default)\n"
+        "  -l, --lex             Lex the specified template into tokens\n"
+        "  -p, --parse           Parse the specified template into an AST\n"
+        "  -c, --compile         Compile the specified template into opcodes\n"
+        "  -V, --version         Print the version\n"
+        "\n"
+        "Input options:\n"
+        "  -t, --template=FILE   The template to operate on\n"
+        "  -D, --data=FILE       The input data file. Supports JSON and YAML.\n"
+        "\n"
+        "Behavior options:\n"
+        "  --flags=FLAGS         The flags to pass to the compiler separated by commas. One or more of:\n"
+        "                        compat, known_helpers_only, string_params, track_ids ,no_escape,\n"
+        "                        ignore_standalone, alternate_decorators, strict, assume_objects,\n"
+        "                        mustache_style_lambdas\n"
+        "  --partial-loader      Specify to enable loading partials dynamically\n"
+        "  --partial-path=DIR    The directory in which to look for partials\n"
+        "  --partial-ext=EXT     The file extension of partials, including the '.'\n"
+        "\n"
+        "The partial loader will concat the partial-path, given partial name in the template,\n"
+        "and the partial-extension to resolve the file from which to load the partial.\n"
+        "\n"
+        "If a FILE is specified as '-', it will be read from STDIN.\n"
+        "\n"
+        "handlebarsc home page: " PACKAGE_URL
+    );
     return 0;
 }
 
 static int do_version(void)
 {
-    fprintf(stderr, "handlebarsc v%s\n", handlebars_version_string());
+    fprintf(stdout,
+        "handlebarsc v%s\n"
+        "Copyright (C) 2020 John Boehr\n"
+        "License LGPLv2.1+: Lesser GNU GPL version 2.1 or later <https://www.gnu.org/licenses/lgpl-2.1.html>.\n"
+        "This is free software: you are free to change and redistribute it.\n"
+        "There is NO WARRANTY, to the extent permitted by law.\n",
+        handlebars_version_string());
     return 0;
 }
 
@@ -436,7 +486,16 @@ static int do_execute(void)
     compiler = handlebars_compiler_ctor(ctx);
     vm = handlebars_vm_ctor(ctx);
     vm->helpers = handlebars_value_ctor(ctx);
-    vm->partials = handlebars_value_ctor(ctx);
+
+    if (enable_partial_loader) {
+        struct handlebars_string *partial_path_str = NULL;
+        struct handlebars_string *partial_extension_str = NULL;
+        partial_path_str = handlebars_string_ctor(ctx, partial_path, strlen(partial_path));
+        partial_extension_str = handlebars_string_ctor(ctx, partial_extension, strlen(partial_extension));
+        vm->partials = handlebars_value_partial_loader_ctor(ctx, partial_path_str, partial_extension_str);
+    } else {
+        vm->partials = handlebars_value_ctor(ctx);
+    }
 
     if( compiler_flags & handlebars_compiler_flag_ignore_standalone ) {
         parser->ignore_standalone = 1;
@@ -456,9 +515,16 @@ static int do_execute(void)
     // Read context
     struct handlebars_value * context = NULL;
     if( input_data_name ) {
+        size_t input_data_name_len = strlen(input_data_name);
         char * context_str = file_get_contents(input_data_name);
-        if( context_str && strlen(context_str) ) {
-            context = handlebars_value_from_json_string(ctx, context_str);
+        if (context_str && strlen(context_str)) {
+            if (input_data_name_len > 5 && (0 == strcmp(input_data_name + input_data_name_len - 5, ".yaml") ||
+                    0 == strcmp(input_data_name + input_data_name_len - 4, ".yml"))) {
+                context = handlebars_value_from_yaml_string(ctx, context_str);
+            } else {
+                // assume json
+                context = handlebars_value_from_json_string(ctx, context_str);
+            }
         }
     }
     if( !context ) {